adjust structures
adjust structure to make it easier to deploy.
