+++ app/static/css/wide.css
