--- models.py
@@ -1,34 +1,0 @@
-#!/usr/bin/env python
-#coding:utf-8
-
-from flask_mongoengine import *
-
-class User(Document):
-	email = StringField(required=True)
-	first_name = StringField(max_length=50)
-	last_name = StringField(max_length=50)
-
-class Post(Document):
-	title = StringField(max_length=120, required=True)
-	author = ReferenceField(User, reverse_delete_rule=CASCADE)
-	tags = ListField(StringField(max_length=30))
-	comments = ListField(EmbeddedDocumentField(Comment))
-
-	meta = {'allow_inheritance': True}
-
-class TextPost(Post):
-	content = StringField()
-
-class ImagePost(Post):
-	image_path = StringField()
-
-class LinkPost(Post):
-	link_url = StringField()
-
-class Comment(EmbeddedDocument):
-	content = StringField()
-	name = StringField(max_length=120)
-
-if __name__ == '__main__':
-	ross = User(email='ross@email.com', first_name='Ross',last_name='Lawley').save()
-
