+++ readme.md
@@ -1,3 +1,8 @@
 # A simple Flask blog app
 
 A simple flask based blog let you write articles with markdown.
+
+## utilities
+
+* generate site from markdown files.
+* configurable pagination.
